refactor: restructure info package
